it_IT language for Animania
Hi Purplicious_Cow! This is the italian language for Animania. Contact me if any new updates come with new stuff to translate, and I'll be happy to deliver! It has reached me that another it_IT language has been done, so if you like this one better, just change this file's name to "it_IT.lang" and delete the other one (or merge them, whatddya think?).

-Pterodattilojr